viixoo_app_engine/viixoo_core/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]  # Add any other build dependencies
build-backend = "setuptools.build_meta"

[project]
name = "viixoo_core" # Use hyphens for package names on PyPI
version = "0.1.1"  # Update your version number
authors = [
    { name = "Vladimir Alvarez Perez", email = "vladimir@viixoo.com" },
]
description = "Microframework based on FastAPI and Odoo for modular and distributed systems in Viixoo."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
keywords = ["fastapi", "odoo", "microservices", "framework", "Viixoo"]  # Add relevant keywords
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "psycopg2-binary",
    "Jinja2",
]

[project.urls]
"Homepage" = "https://github.com/vladimir881002/viixoo_app_engine"
"Bug Tracker" = "https://github.com/vladimir881002/viixoo_app_engine/issues"  # (Optional)

[project.optional-dependencies]  # (Optional) Define extra dependencies
dev = ["pytest", "pre-commit", "black", "isort"]  # development dependencies

[project.entry-points."console_scripts"]  # Note the quotes around "console_scripts"
viixoo_core_run = "viixoo_core.app:app"  # Your entry point
viixoo_migrate = "viixoo_core.migrations:Migration.run"
viixoo_convert = "viixoo_core.odoo_converter.converter:main"
# Add other entry points here if needed

[tool.setuptools.packages.find]
where = ["."]
